Take samples for linearization only during operation. Adapt to new signal I/O interface

+++ src/output.c
@@ -35,7 +35,7 @@
 struct _OutputData
 {
   DECLARE_MODULE_INTERFACE_REF( SIGNAL_IO_INTERFACE );
-  int deviceID;
+  long int deviceID;
   unsigned int channel;
 };
